refactor(reader): split oversized reader screens into section files
Each reader screen had its sheets, dialogs, page-render modes, and control
overlays inline. Extracted those private/internal composables to sibling
files in the same package so the orchestrator screens stay below the
800-LOC guardrail.

- MediaBookReader.kt: 1420 -> 539 LOC (15 sibling files extracted)
- EnhancedPDFReaderScreen.kt: 1273 -> 751 LOC (7 sibling files)
- MediaReaderScreen.kt: 1020 -> 484 LOC (5 sibling files)
- AdvancedComicReader.kt: 967 -> 475 LOC (6 sibling files)
- EnhancedEReaderScreen.kt: 817 -> 586 LOC (3 sibling files)
- DocumentReaderComponents.kt: 1133 -> 500 LOC (3 sibling files)
