+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/MediaReaderScreen.kt
@@ -212,74 +212,6 @@ fun MediaReaderScreen(
 // READER CONTENT
 // =============================================================================
 
-@Composable
-private fun ReaderContent(
-    content: ReaderContent,
-    theme: ReaderTheme,
-    typography: ReaderTypography,
-    isComic: Boolean,
-    currentPage: Int,
-    degradedWarning: String?
-) {
-    if (isComic) {
-        // Comic/manga page view
-        ComicPageView(
-            imageUrl = content.imageUrls.getOrNull(currentPage - 1) ?: content.imageUrl,
-            modifier = Modifier.fillMaxSize()
-        )
-    } else {
-        if (!content.htmlContent.isNullOrBlank()) {
-            HtmlReaderContent(
-                htmlContent = content.htmlContent,
-                theme = theme,
-                typography = typography
-            )
-        } else {
-            // Text content
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(horizontal = typography.marginHorizontal.dp),
-                contentPadding = PaddingValues(vertical = typography.marginVertical.dp)
-            ) {
-                if (!degradedWarning.isNullOrBlank()) {
-                    item {
-                        Surface(
-                            shape = RoundedCornerShape(999.dp),
-                            color = MediaColors.Warning.copy(alpha = 0.2f),
-                            tonalElevation = 0.dp,
-                            modifier = Modifier.padding(bottom = 12.dp)
-                        ) {
-                            Text(
-                                text = "⚠ $degradedWarning",
-                                style = MaterialTheme.typography.labelMedium,
-                                color = MediaColors.Warning,
-                                modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp)
-                            )
-                        }
-                    }
-                }
-                item {
-                    Text(
-                        text = content.text,
-                        color = theme.textColor,
-                        fontSize = typography.fontSize.sp,
-                        lineHeight = (typography.fontSize * typography.lineHeight).sp,
-                        fontFamily = typography.fontFamily,
-                        letterSpacing = typography.letterSpacing.sp,
-                        textAlign = when (typography.textAlign) {
-                            TextAlignment.LEFT -> TextAlign.Start
-                            TextAlignment.CENTER -> TextAlign.Center
-                            TextAlignment.RIGHT -> TextAlign.End
-                            TextAlignment.JUSTIFY -> TextAlign.Justify
-                        },
-                        style = MaterialTheme.typography.bodyLarge
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun HtmlReaderContent(
@@ -410,489 +342,21 @@ private fun ComicPageView(
 // TOP BAR
 // =============================================================================
 
-@Composable
-private fun ReaderTopBar(
-    title: String,
-    chapter: String?,
-    isBookmarked: Boolean,
-    isTtsActive: Boolean,
-    onClose: () -> Unit,
-    onBookmark: () -> Unit,
-    onToc: () -> Unit,
-    onSearch: () -> Unit,
-    onTts: () -> Unit,
-    onSettings: () -> Unit
-) {
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        color = MediaColors.Background.copy(alpha = 0.95f),
-        tonalElevation = MediaElevation.SM
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .statusBarsPadding()
-                .padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.SM),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Close button
-            IconButton(onClick = onClose) {
-                Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = "Close",
-                    tint = MediaColors.TextPrimary
-                )
-            }
-            
-            // Title & Chapter
-            Column(
-                modifier = Modifier.weight(1f),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = title,
-                    style = MediaTypography.BodyMedium,
-                    color = MediaColors.TextPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = FontWeight.Medium
-                )
-                chapter?.let {
-                    Text(
-                        text = it,
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextSecondary,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-            }
-            
-            // Action buttons
-            Row {
-                IconButton(onClick = onBookmark) {
-                    Icon(
-                        imageVector = if (isBookmarked) Icons.Filled.Bookmark else Icons.Outlined.BookmarkBorder,
-                        contentDescription = "Bookmark",
-                        tint = if (isBookmarked) MediaColors.AccentPrimary else MediaColors.TextSecondary
-                    )
-                }
-                
-                IconButton(onClick = onTts) {
-                    Icon(
-                        imageVector = if (isTtsActive) Icons.Filled.VolumeUp else Icons.Outlined.VolumeUp,
-                        contentDescription = "Text to Speech",
-                        tint = if (isTtsActive) MediaColors.AccentPrimary else MediaColors.TextSecondary
-                    )
-                }
-                
-                IconButton(onClick = onToc) {
-                    Icon(
-                        imageVector = Icons.Default.List,
-                        contentDescription = "Table of Contents",
-                        tint = MediaColors.TextSecondary
-                    )
-                }
-                
-                IconButton(onClick = onSearch) {
-                    Icon(
-                        imageVector = Icons.Default.Search,
-                        contentDescription = "Search",
-                        tint = MediaColors.TextSecondary
-                    )
-                }
-                
-                IconButton(onClick = onSettings) {
-                    Icon(
-                        imageVector = Icons.Default.Settings,
-                        contentDescription = "Settings",
-                        tint = MediaColors.TextSecondary
-                    )
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // BOTTOM BAR
 // =============================================================================
 
-@Composable
-private fun ReaderBottomBar(
-    currentPage: Int,
-    totalPages: Int,
-    chapterProgress: Float,
-    currentChapterIndex: Int,
-    totalChapters: Int,
-    onPageChange: (Int) -> Unit,
-    onPreviousChapter: () -> Unit,
-    onNextChapter: () -> Unit,
-    theme: ReaderTheme
-) {
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        color = MediaColors.Background.copy(alpha = 0.95f),
-        tonalElevation = MediaElevation.SM
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .navigationBarsPadding()
-                .padding(MediaSpacing.MD)
-        ) {
-            // Page slider
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = currentPage.toString(),
-                    style = MediaTypography.LabelMedium,
-                    color = MediaColors.TextSecondary,
-                    modifier = Modifier.width(40.dp)
-                )
-                
-                Slider(
-                    value = currentPage.toFloat(),
-                    onValueChange = { onPageChange(it.toInt()) },
-                    valueRange = 1f..totalPages.toFloat(),
-                    modifier = Modifier.weight(1f),
-                    colors = SliderDefaults.colors(
-                        thumbColor = MediaColors.AccentPrimary,
-                        activeTrackColor = MediaColors.AccentPrimary,
-                        inactiveTrackColor = MediaColors.ProgressBackground
-                    )
-                )
-                
-                Text(
-                    text = totalPages.toString(),
-                    style = MediaTypography.LabelMedium,
-                    color = MediaColors.TextSecondary,
-                    modifier = Modifier.width(40.dp),
-                    textAlign = TextAlign.End
-                )
-            }
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.SM))
-            
-            // Chapter navigation
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                // Previous chapter
-                TextButton(
-                    onClick = onPreviousChapter,
-                    enabled = currentChapterIndex > 0
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.ChevronLeft,
-                        contentDescription = "Media image",
-                        modifier = Modifier.size(20.dp)
-                    )
-                    Text("Prev")
-                }
-                
-                // Chapter info
-                Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                    Text(
-                        text = "Chapter ${currentChapterIndex + 1} of $totalChapters",
-                        style = MediaTypography.LabelMedium,
-                        color = MediaColors.TextSecondary
-                    )
-                    LinearProgressIndicator(
-                        progress = { chapterProgress },
-                        modifier = Modifier
-                            .width(100.dp)
-                            .height(4.dp)
-                            .clip(RoundedCornerShape(MediaCorners.Full)),
-                        color = MediaColors.AccentPrimary,
-                        trackColor = MediaColors.ProgressBackground
-                    )
-                }
-                
-                // Next chapter
-                TextButton(
-                    onClick = onNextChapter,
-                    enabled = currentChapterIndex < totalChapters - 1
-                ) {
-                    Text("Next")
-                    Icon(
-                        imageVector = Icons.Default.ChevronRight,
-                        contentDescription = "Media image",
-                        modifier = Modifier.size(20.dp)
-                    )
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // TABLE OF CONTENTS SHEET
 // =============================================================================
 
-@Composable
-private fun TocSheet(
-    chapters: List<ChapterInfo>,
-    currentChapterIndex: Int,
-    onChapterSelect: (Int) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Table of Contents",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.MD)
-        )
-        
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(max = 400.dp)
-        ) {
-            itemsIndexed(chapters) { index, chapter ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { onChapterSelect(index) }
-                        .background(
-                            if (index == currentChapterIndex)
-                                MediaColors.AccentPrimary.copy(alpha = 0.1f)
-                            else
-                                Color.Transparent
-                        )
-                        .padding(MediaSpacing.MD),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = chapter.title,
-                            style = MediaTypography.BodyMedium,
-                            color = if (index == currentChapterIndex)
-                                MediaColors.AccentPrimary
-                            else
-                                MediaColors.TextPrimary,
-                            fontWeight = if (index == currentChapterIndex)
-                                FontWeight.SemiBold
-                            else
-                                FontWeight.Normal
-                        )
-                        
-                        if (chapter.progress > 0) {
-                            Spacer(modifier = Modifier.height(MediaSpacing.XS))
-                            LinearProgressIndicator(
-                                progress = { chapter.progress },
-                                modifier = Modifier
-                                    .fillMaxWidth(0.5f)
-                                    .height(2.dp)
-                                    .clip(RoundedCornerShape(MediaCorners.Full)),
-                                color = MediaColors.AccentPrimary,
-                                trackColor = MediaColors.ProgressBackground
-                            )
-                        }
-                    }
-                    
-                    if (index == currentChapterIndex) {
-                        Icon(
-                            imageVector = Icons.Default.PlayArrow,
-                            contentDescription = "Current",
-                            tint = MediaColors.AccentPrimary,
-                            modifier = Modifier.size(20.dp)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // SETTINGS SHEET
 // =============================================================================
 
-@Composable
-private fun ReaderSettingsSheet(
-    theme: ReaderTheme,
-    typography: ReaderTypography,
-    onThemeChange: (ReaderTheme) -> Unit,
-    onTypographyChange: (ReaderTypography) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Reading Settings",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.LG)
-        )
-        
-        // Theme selector
-        Text(
-            text = "Theme",
-            style = MediaTypography.LabelMedium,
-            color = MediaColors.TextSecondary,
-            modifier = Modifier.padding(bottom = MediaSpacing.SM)
-        )
-        
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-        ) {
-            ReaderTheme.presets.forEach { preset ->
-                ThemeButton(
-                    theme = preset,
-                    isSelected = theme.name == preset.name,
-                    onClick = { onThemeChange(preset) },
-                    modifier = Modifier.weight(1f)
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.LG))
-        
-        // Font size
-        Text(
-            text = "Font Size",
-            style = MediaTypography.LabelMedium,
-            color = MediaColors.TextSecondary,
-            modifier = Modifier.padding(bottom = MediaSpacing.SM)
-        )
-        
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            IconButton(onClick = {
-                if (typography.fontSize > 12) {
-                    onTypographyChange(typography.copy(fontSize = typography.fontSize - 1))
-                }
-            }) {
-                Icon(
-                    imageVector = Icons.Default.TextDecrease,
-                    contentDescription = "Decrease",
-                    tint = MediaColors.TextSecondary
-                )
-            }
-            
-            Text(
-                text = "${typography.fontSize}sp",
-                style = MediaTypography.BodyMedium,
-                color = MediaColors.TextPrimary
-            )
-            
-            IconButton(onClick = {
-                if (typography.fontSize < 32) {
-                    onTypographyChange(typography.copy(fontSize = typography.fontSize + 1))
-                }
-            }) {
-                Icon(
-                    imageVector = Icons.Default.TextIncrease,
-                    contentDescription = "Increase",
-                    tint = MediaColors.TextSecondary
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.LG))
-        
-        // Line height
-        Text(
-            text = "Line Spacing",
-            style = MediaTypography.LabelMedium,
-            color = MediaColors.TextSecondary,
-            modifier = Modifier.padding(bottom = MediaSpacing.SM)
-        )
-        
-        Slider(
-            value = typography.lineHeight,
-            onValueChange = { onTypographyChange(typography.copy(lineHeight = it)) },
-            valueRange = 1.0f..2.5f,
-            steps = 5,
-            colors = SliderDefaults.colors(
-                thumbColor = MediaColors.AccentPrimary,
-                activeTrackColor = MediaColors.AccentPrimary
-            )
-        )
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.LG))
-        
-        // Margins
-        Text(
-            text = "Margins",
-            style = MediaTypography.LabelMedium,
-            color = MediaColors.TextSecondary,
-            modifier = Modifier.padding(bottom = MediaSpacing.SM)
-        )
-        
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-        ) {
-            listOf(8, 16, 24, 32).forEach { margin ->
-                FilterChip(
-                    selected = typography.marginHorizontal == margin,
-                    onClick = { onTypographyChange(typography.copy(marginHorizontal = margin)) },
-                    label = { Text("${margin}dp") },
-                    colors = FilterChipDefaults.filterChipColors(
-                        selectedContainerColor = MediaColors.AccentPrimary.copy(alpha = 0.2f),
-                        selectedLabelColor = MediaColors.AccentPrimary
-                    )
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.LG))
-        
-        // Text alignment
-        Text(
-            text = "Text Align",
-            style = MediaTypography.LabelMedium,
-            color = MediaColors.TextSecondary,
-            modifier = Modifier.padding(bottom = MediaSpacing.SM)
-        )
-        
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-        ) {
-            TextAlignment.entries.forEach { align ->
-                FilterChip(
-                    selected = typography.textAlign == align,
-                    onClick = { onTypographyChange(typography.copy(textAlign = align)) },
-                    label = { 
-                        Icon(
-                            imageVector = when (align) {
-                                TextAlignment.LEFT -> Icons.Default.FormatAlignLeft
-                                TextAlignment.CENTER -> Icons.Default.FormatAlignCenter
-                                TextAlignment.RIGHT -> Icons.Default.FormatAlignRight
-                                TextAlignment.JUSTIFY -> Icons.Default.FormatAlignJustify
-                            },
-                            contentDescription = align.name,
-                            modifier = Modifier.size(18.dp)
-                        )
-                    },
-                    colors = FilterChipDefaults.filterChipColors(
-                        selectedContainerColor = MediaColors.AccentPrimary.copy(alpha = 0.2f),
-                        selectedLabelColor = MediaColors.AccentPrimary
-                    )
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.XL))
-    }
-}
 
 @Composable
 private fun ThemeButton(
