+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/BookmarksList.kt
@@ -1,0 +1,77 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun BookmarksList(
+    bookmarks: List<BookmarkInfo>,
+    onBookmarkClick: (BookmarkInfo) -> Unit
+) {
+    if (bookmarks.isEmpty()) {
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(MediaSpacing.XL),
+            contentAlignment = Alignment.Center
+        ) {
+            Text("No bookmarks yet", color = MediaColors.TextSecondary)
+        }
+    } else {
+        LazyColumn {
+            items(items = bookmarks, key = { it.id }) { bookmark ->
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .clickable { onBookmarkClick(bookmark) }
+                        .padding(MediaSpacing.MD),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Icon(Icons.Default.Bookmark, null, tint = MediaColors.AccentPrimary)
+                    Spacer(Modifier.width(MediaSpacing.MD))
+                    Column(modifier = Modifier.weight(1f)) {
+                        Text(
+                            "Page ${bookmark.pageNumber + 1}",
+                            style = MediaTypography.BodyMedium,
+                            color = MediaColors.TextPrimary
+                        )
+                        Text(
+                            bookmark.dateAdded,
+                            style = MediaTypography.LabelSmall,
+                            color = MediaColors.TextTertiary
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
