+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/BookmarksSheet.kt
@@ -1,0 +1,77 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun BookmarksSheet(
+    bookmarks: List<BookmarkInfo>,
+    highlights: List<HighlightInfo>,
+    annotations: List<AnnotationInfo>,
+    onBookmarkClick: (BookmarkInfo) -> Unit,
+    onDismiss: () -> Unit
+) {
+    var selectedTab by remember { mutableStateOf(0) }
+    val tabs = listOf("Bookmarks", "Highlights", "Notes")
+    
+    ModalBottomSheet(
+        onDismissRequest = onDismiss,
+        containerColor = MediaColors.BackgroundElevated
+    ) {
+        Column(modifier = Modifier.fillMaxWidth()) {
+            TabRow(
+                selectedTabIndex = selectedTab,
+                containerColor = Color.Transparent
+            ) {
+                tabs.forEachIndexed { index, title ->
+                    Tab(
+                        selected = selectedTab == index,
+                        onClick = { selectedTab = index },
+                        text = {
+                            Text(
+                                title,
+                                color = if (selectedTab == index) MediaColors.AccentPrimary else MediaColors.TextSecondary
+                            )
+                        }
+                    )
+                }
+            }
+            
+            when (selectedTab) {
+                0 -> BookmarksList(bookmarks, onBookmarkClick)
+                1 -> HighlightsList(highlights)
+                2 -> AnnotationsList(annotations)
+            }
+        }
+    }
+}
