+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/MediaBookReader.kt
@@ -257,373 +257,14 @@ fun MediaBookReader(
 // CONTENT MODES
 // =============================================================================
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun PaginatedContent(
-    pagerState: PagerState,
-    pages: List<BookPage>,
-    settings: ReaderSettings,
-    highlights: List<HighlightInfo>,
-    pageAnimation: PageAnimation
-) {
-    HorizontalPager(
-        state = pagerState,
-        modifier = Modifier.fillMaxSize(),
-        beyondViewportPageCount = 1,
-        pageSpacing = 0.dp
-    ) { pageIndex ->
-        val page = pages.getOrNull(pageIndex)
-        
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(
-                    horizontal = settings.marginHorizontal.dp,
-                    vertical = settings.marginVertical.dp
-                )
-                .graphicsLayer {
-                    // Apply page turn animation
-                    when (pageAnimation) {
-                        PageAnimation.SLIDE -> {
-                            val pageOffset = (pagerState.currentPage - pageIndex) + pagerState.currentPageOffsetFraction
-                            translationX = pageOffset * size.width
-                        }
-                        PageAnimation.CURL -> {
-                            val pageOffset = (pagerState.currentPage - pageIndex) + pagerState.currentPageOffsetFraction
-                            rotationY = pageOffset * -30f
-                            cameraDistance = 8 * density
-                        }
-                        PageAnimation.FADE -> {
-                            val pageOffset = (pagerState.currentPage - pageIndex) + pagerState.currentPageOffsetFraction
-                            alpha = 1f - kotlin.math.abs(pageOffset)
-                        }
-                        PageAnimation.NONE -> { }
-                    }
-                }
-        ) {
-            if (page != null) {
-                BookPageContent(
-                    page = page,
-                    settings = settings,
-                    highlights = highlights.filter { it.pageNumber == pageIndex }
-                )
-            }
-        }
-    }
-}
 
-@Composable
-private fun ScrollContent(
-    content: String,
-    settings: ReaderSettings,
-    highlights: List<HighlightInfo>,
-    onScrollPosition: (Float) -> Unit
-) {
-    val scrollState = rememberScrollState()
-    
-    LaunchedEffect(scrollState.value) {
-        if (scrollState.maxValue > 0) {
-            onScrollPosition(scrollState.value.toFloat() / scrollState.maxValue)
-        }
-    }
-    
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .verticalScroll(scrollState)
-            .padding(
-                horizontal = settings.marginHorizontal.dp,
-                vertical = settings.marginVertical.dp
-            )
-    ) {
-        Text(
-            text = content,
-            style = TextStyle(
-                fontFamily = getFontFamily(settings.fontFamily),
-                fontSize = settings.fontSize.sp,
-                lineHeight = (settings.fontSize * settings.lineHeight).sp,
-                color = settings.theme.textColor,
-                textAlign = when (settings.textAlign) {
-                    ReaderTextAlign.LEFT -> TextAlign.Start
-                    ReaderTextAlign.CENTER -> TextAlign.Center
-                    ReaderTextAlign.RIGHT -> TextAlign.End
-                    ReaderTextAlign.JUSTIFIED -> TextAlign.Justify
-                }
-            )
-        )
-    }
-}
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun TwoPageContent(
-    pagerState: PagerState,
-    pages: List<BookPage>,
-    settings: ReaderSettings
-) {
-    val configuration = LocalConfiguration.current
-    val isLandscape = configuration.screenWidthDp > configuration.screenHeightDp
-    
-    if (isLandscape) {
-        HorizontalPager(
-            state = pagerState,
-            modifier = Modifier.fillMaxSize()
-        ) { spreadIndex ->
-            Row(modifier = Modifier.fillMaxSize()) {
-                // Left page
-                val leftPageIndex = spreadIndex * 2
-                Box(
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxHeight()
-                        .padding(start = settings.marginHorizontal.dp, end = 8.dp)
-                ) {
-                    pages.getOrNull(leftPageIndex)?.let { page ->
-                        BookPageContent(
-                            page = page,
-                            settings = settings,
-                            highlights = emptyList()
-                        )
-                    }
-                }
-                
-                // Divider
-                VerticalDivider(
-                    color = settings.theme.textColor.copy(alpha = 0.1f),
-                    modifier = Modifier.fillMaxHeight()
-                )
-                
-                // Right page
-                val rightPageIndex = leftPageIndex + 1
-                Box(
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxHeight()
-                        .padding(start = 8.dp, end = settings.marginHorizontal.dp)
-                ) {
-                    pages.getOrNull(rightPageIndex)?.let { page ->
-                        BookPageContent(
-                            page = page,
-                            settings = settings,
-                            highlights = emptyList()
-                        )
-                    }
-                }
-            }
-        }
-    } else {
-        // Fall back to single page in portrait
-        PaginatedContent(
-            pagerState = pagerState,
-            pages = pages,
-            settings = settings,
-            highlights = emptyList(),
-            pageAnimation = settings.pageAnimation
-        )
-    }
-}
 
-@Composable
-private fun BookPageContent(
-    page: BookPage,
-    settings: ReaderSettings,
-    highlights: List<HighlightInfo>
-) {
-    Column(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        // Main text content
-        Text(
-            text = page.content,
-            style = TextStyle(
-                fontFamily = getFontFamily(settings.fontFamily),
-                fontSize = settings.fontSize.sp,
-                lineHeight = (settings.fontSize * settings.lineHeight).sp,
-                color = settings.theme.textColor,
-                textAlign = when (settings.textAlign) {
-                    ReaderTextAlign.LEFT -> TextAlign.Start
-                    ReaderTextAlign.CENTER -> TextAlign.Center
-                    ReaderTextAlign.RIGHT -> TextAlign.End
-                    ReaderTextAlign.JUSTIFIED -> TextAlign.Justify
-                }
-            ),
-            modifier = Modifier.weight(1f)
-        )
-        
-        // Page number
-        if (settings.showPageNumber) {
-            Text(
-                text = "${page.pageNumber + 1}",
-                style = TextStyle(
-                    fontSize = 12.sp,
-                    color = settings.theme.textColor.copy(alpha = 0.5f),
-                    textAlign = TextAlign.Center
-                ),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = MediaSpacing.SM)
-            )
-        }
-    }
-}
 
 // =============================================================================
 // CONTROLS OVERLAY
 // =============================================================================
 
-@Composable
-private fun ReaderControlsOverlay(
-    state: BookReaderState,
-    onClose: () -> Unit,
-    onShowSettings: () -> Unit,
-    onShowTOC: () -> Unit,
-    onShowSearch: () -> Unit,
-    onShowBookmarks: () -> Unit,
-    onBookmarkToggle: () -> Unit,
-    onTTSToggle: () -> Unit,
-    onPageSelect: (Int) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.4f))
-    ) {
-        // Top bar
-        Surface(
-            color = MediaColors.BackgroundElevated.copy(alpha = 0.95f),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .statusBarsPadding()
-                    .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                IconButton(onClick = onClose) {
-                    Icon(Icons.Default.Close, "Close", tint = MediaColors.TextPrimary)
-                }
-                
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(
-                        text = state.bookTitle,
-                        style = MediaTypography.TitleSmall,
-                        color = MediaColors.TextPrimary,
-                        maxLines = 1
-                    )
-                    state.currentChapter?.title?.let { chapter ->
-                        Text(
-                            text = chapter,
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextSecondary,
-                            maxLines = 1
-                        )
-                    }
-                }
-                
-                IconButton(onClick = onBookmarkToggle) {
-                    Icon(
-                        if (state.isCurrentPageBookmarked) Icons.Default.Bookmark else Icons.Outlined.BookmarkBorder,
-                        "Bookmark",
-                        tint = if (state.isCurrentPageBookmarked) MediaColors.AccentPrimary else MediaColors.TextSecondary
-                    )
-                }
-                
-                IconButton(onClick = onTTSToggle) {
-                    Icon(
-                        if (state.isTTSPlaying) Icons.Default.VolumeUp else Icons.Outlined.VolumeUp,
-                        "Text to Speech",
-                        tint = if (state.isTTSPlaying) MediaColors.AccentPrimary else MediaColors.TextSecondary
-                    )
-                }
-            }
-        }
-        
-        Spacer(modifier = Modifier.weight(1f))
-        
-        // Bottom bar
-        Surface(
-            color = MediaColors.BackgroundElevated.copy(alpha = 0.95f),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .navigationBarsPadding()
-                    .padding(MediaSpacing.MD)
-            ) {
-                // Progress slider
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Text(
-                        text = "${state.currentPage + 1}",
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextSecondary
-                    )
-                    
-                    Slider(
-                        value = state.currentPage.toFloat(),
-                        onValueChange = { onPageSelect(it.toInt()) },
-                        valueRange = 0f..(state.totalPages - 1).toFloat(),
-                        modifier = Modifier
-                            .weight(1f)
-                            .padding(horizontal = MediaSpacing.SM),
-                        colors = SliderDefaults.colors(
-                            thumbColor = MediaColors.AccentPrimary,
-                            activeTrackColor = MediaColors.AccentPrimary
-                        )
-                    )
-                    
-                    Text(
-                        text = "${state.totalPages}",
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextSecondary
-                    )
-                }
-                
-                // Progress percentage
-                Text(
-                    text = "${((state.currentPage + 1).toFloat() / state.totalPages * 100).toInt()}% complete",
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextTertiary,
-                    modifier = Modifier.fillMaxWidth(),
-                    textAlign = TextAlign.Center
-                )
-                
-                Spacer(modifier = Modifier.height(MediaSpacing.MD))
-                
-                // Action buttons
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-                    ReaderActionButton(
-                        icon = Icons.Default.List,
-                        label = "Contents",
-                        onClick = onShowTOC
-                    )
-                    ReaderActionButton(
-                        icon = Icons.Default.Search,
-                        label = "Search",
-                        onClick = onShowSearch
-                    )
-                    ReaderActionButton(
-                        icon = Icons.Default.Bookmarks,
-                        label = "Bookmarks",
-                        onClick = onShowBookmarks
-                    )
-                    ReaderActionButton(
-                        icon = Icons.Default.Settings,
-                        label = "Settings",
-                        onClick = onShowSettings
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun ReaderActionButton(
@@ -681,165 +322,6 @@ private fun TTSIndicator(modifier: Modifier = Modifier) {
 // SETTINGS SHEET
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ReaderSettingsSheet(
-    settings: ReaderSettings,
-    onSettingsChange: (ReaderSettings) -> Unit,
-    onDismiss: () -> Unit
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        containerColor = MediaColors.BackgroundElevated
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.MD)
-                .verticalScroll(rememberScrollState())
-        ) {
-            Text(
-                "Reader Settings",
-                style = MediaTypography.TitleMedium,
-                color = MediaColors.TextPrimary,
-                modifier = Modifier.padding(bottom = MediaSpacing.MD)
-            )
-            
-            // Font size
-            SettingSection(title = "Font Size") {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    IconButton(onClick = {
-                        if (settings.fontSize > 12) {
-                            onSettingsChange(settings.copy(fontSize = settings.fontSize - 2))
-                        }
-                    }) {
-                        Icon(Icons.Default.Remove, "Decrease", tint = MediaColors.TextSecondary)
-                    }
-                    
-                    Text(
-                        "${settings.fontSize.toInt()}",
-                        style = MediaTypography.TitleMedium,
-                        color = MediaColors.TextPrimary,
-                        modifier = Modifier.weight(1f),
-                        textAlign = TextAlign.Center
-                    )
-                    
-                    IconButton(onClick = {
-                        if (settings.fontSize < 32) {
-                            onSettingsChange(settings.copy(fontSize = settings.fontSize + 2))
-                        }
-                    }) {
-                        Icon(Icons.Default.Add, "Increase", tint = MediaColors.TextSecondary)
-                    }
-                }
-            }
-            
-            // Font family
-            SettingSection(title = "Font") {
-                LazyRow(
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                ) {
-                    items(ReaderFontFamily.entries.toList()) { font ->
-                        FilterChip(
-                            selected = settings.fontFamily == font,
-                            onClick = { onSettingsChange(settings.copy(fontFamily = font)) },
-                            label = { Text(font.displayName) }
-                        )
-                    }
-                }
-            }
-            
-            // Line height
-            SettingSection(title = "Line Height") {
-                Slider(
-                    value = settings.lineHeight,
-                    onValueChange = { onSettingsChange(settings.copy(lineHeight = it)) },
-                    valueRange = 1.2f..2.5f,
-                    colors = SliderDefaults.colors(
-                        thumbColor = MediaColors.AccentPrimary,
-                        activeTrackColor = MediaColors.AccentPrimary
-                    )
-                )
-            }
-            
-            // Theme presets
-            SettingSection(title = "Theme") {
-                LazyRow(
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                ) {
-                    items(ReaderThemePreset.entries.toList()) { preset ->
-                        ThemePresetChip(
-                            preset = preset,
-                            isSelected = settings.themePreset == preset,
-                            onClick = { 
-                                onSettingsChange(settings.copy(
-                                    themePreset = preset,
-                                    theme = preset.toTheme()
-                                ))
-                            }
-                        )
-                    }
-                }
-            }
-            
-            // Reading mode
-            SettingSection(title = "Reading Mode") {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                ) {
-                    ReadingMode.entries.forEach { mode ->
-                        FilterChip(
-                            selected = settings.readingMode == mode,
-                            onClick = { onSettingsChange(settings.copy(readingMode = mode)) },
-                            label = { Text(mode.displayName) },
-                            modifier = Modifier.weight(1f)
-                        )
-                    }
-                }
-            }
-            
-            // Page animation
-            SettingSection(title = "Page Animation") {
-                LazyRow(
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                ) {
-                    items(PageAnimation.entries.toList()) { animation ->
-                        FilterChip(
-                            selected = settings.pageAnimation == animation,
-                            onClick = { onSettingsChange(settings.copy(pageAnimation = animation)) },
-                            label = { Text(animation.displayName) }
-                        )
-                    }
-                }
-            }
-            
-            // Toggles
-            SettingSection(title = "Display") {
-                SettingToggle(
-                    title = "Show Page Number",
-                    checked = settings.showPageNumber,
-                    onCheckedChange = { onSettingsChange(settings.copy(showPageNumber = it)) }
-                )
-                SettingToggle(
-                    title = "Keep Screen On",
-                    checked = settings.keepScreenOn,
-                    onCheckedChange = { onSettingsChange(settings.copy(keepScreenOn = it)) }
-                )
-                SettingToggle(
-                    title = "Volume Page Turn",
-                    checked = settings.volumePageTurn,
-                    onCheckedChange = { onSettingsChange(settings.copy(volumePageTurn = it)) }
-                )
-            }
-            
-            Spacer(Modifier.height(MediaSpacing.XL))
-        }
-    }
-}
 
 @Composable
 private fun SettingSection(
@@ -916,137 +398,12 @@ private fun ThemePresetChip(
 // TABLE OF CONTENTS
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun TableOfContentsSheet(
-    chapters: List<BookChapter>,
-    currentChapter: BookChapter?,
-    onChapterSelect: (BookChapter) -> Unit,
-    onDismiss: () -> Unit
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        containerColor = MediaColors.BackgroundElevated
-    ) {
-        Column(modifier = Modifier.fillMaxWidth()) {
-            Text(
-                "Table of Contents",
-                style = MediaTypography.TitleMedium,
-                color = MediaColors.TextPrimary,
-                modifier = Modifier.padding(MediaSpacing.MD)
-            )
-            
-            LazyColumn {
-                items(items = chapters, key = { it.id }) { chapter ->
-                    ChapterItem(
-                        chapter = chapter,
-                        isCurrent = chapter.id == currentChapter?.id,
-                        onClick = { onChapterSelect(chapter) }
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun ChapterItem(
-    chapter: BookChapter,
-    isCurrent: Boolean,
-    onClick: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .background(if (isCurrent) MediaColors.AccentPrimary.copy(alpha = 0.1f) else Color.Transparent)
-            .padding(
-                start = (MediaSpacing.MD.value + chapter.level * 16).dp,
-                end = MediaSpacing.MD,
-                top = MediaSpacing.SM,
-                bottom = MediaSpacing.SM
-            ),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        if (isCurrent) {
-            Icon(
-                Icons.Default.PlayArrow,
-                null,
-                tint = MediaColors.AccentPrimary,
-                modifier = Modifier.size(16.dp)
-            )
-            Spacer(Modifier.width(MediaSpacing.SM))
-        }
-        
-        Text(
-            chapter.title,
-            style = MediaTypography.BodyMedium,
-            color = if (isCurrent) MediaColors.AccentPrimary else MediaColors.TextPrimary,
-            fontWeight = if (isCurrent) FontWeight.Medium else FontWeight.Normal
-        )
-    }
-}
 
 // =============================================================================
 // SEARCH
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SearchInBookSheet(
-    searchResults: List<ReaderSearchResult>,
-    onSearch: (String) -> Unit,
-    onResultClick: (ReaderSearchResult) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var query by remember { mutableStateOf("") }
-    
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        containerColor = MediaColors.BackgroundElevated
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.MD)
-        ) {
-            OutlinedTextField(
-                value = query,
-                onValueChange = {
-                    query = it
-                    onSearch(it)
-                },
-                placeholder = { Text("Search in book...", color = MediaColors.TextTertiary) },
-                leadingIcon = { Icon(Icons.Default.Search, null, tint = MediaColors.TextSecondary) },
-                modifier = Modifier.fillMaxWidth(),
-                colors = OutlinedTextFieldDefaults.colors(
-                    focusedBorderColor = MediaColors.AccentPrimary,
-                    unfocusedBorderColor = MediaColors.Border
-                )
-            )
-            
-            Spacer(Modifier.height(MediaSpacing.MD))
-            
-            if (searchResults.isEmpty() && query.isNotEmpty()) {
-                Text(
-                    "No results found",
-                    style = MediaTypography.BodyMedium,
-                    color = MediaColors.TextSecondary,
-                    modifier = Modifier.padding(MediaSpacing.MD)
-                )
-            } else {
-                LazyColumn {
-                    items(items = searchResults, key = { "${it.pageNumber}-${it.excerpt.hashCode()}" }) { result ->
-                        SearchResultItem(
-                            result = result,
-                            onClick = { onResultClick(result) }
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun SearchResultItem(
@@ -1080,180 +437,9 @@ private fun SearchResultItem(
 // BOOKMARKS
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun BookmarksSheet(
-    bookmarks: List<BookmarkInfo>,
-    highlights: List<HighlightInfo>,
-    annotations: List<AnnotationInfo>,
-    onBookmarkClick: (BookmarkInfo) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var selectedTab by remember { mutableStateOf(0) }
-    val tabs = listOf("Bookmarks", "Highlights", "Notes")
-    
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        containerColor = MediaColors.BackgroundElevated
-    ) {
-        Column(modifier = Modifier.fillMaxWidth()) {
-            TabRow(
-                selectedTabIndex = selectedTab,
-                containerColor = Color.Transparent
-            ) {
-                tabs.forEachIndexed { index, title ->
-                    Tab(
-                        selected = selectedTab == index,
-                        onClick = { selectedTab = index },
-                        text = {
-                            Text(
-                                title,
-                                color = if (selectedTab == index) MediaColors.AccentPrimary else MediaColors.TextSecondary
-                            )
-                        }
-                    )
-                }
-            }
-            
-            when (selectedTab) {
-                0 -> BookmarksList(bookmarks, onBookmarkClick)
-                1 -> HighlightsList(highlights)
-                2 -> AnnotationsList(annotations)
-            }
-        }
-    }
-}
 
-@Composable
-private fun BookmarksList(
-    bookmarks: List<BookmarkInfo>,
-    onBookmarkClick: (BookmarkInfo) -> Unit
-) {
-    if (bookmarks.isEmpty()) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.XL),
-            contentAlignment = Alignment.Center
-        ) {
-            Text("No bookmarks yet", color = MediaColors.TextSecondary)
-        }
-    } else {
-        LazyColumn {
-            items(items = bookmarks, key = { it.id }) { bookmark ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { onBookmarkClick(bookmark) }
-                        .padding(MediaSpacing.MD),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(Icons.Default.Bookmark, null, tint = MediaColors.AccentPrimary)
-                    Spacer(Modifier.width(MediaSpacing.MD))
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            "Page ${bookmark.pageNumber + 1}",
-                            style = MediaTypography.BodyMedium,
-                            color = MediaColors.TextPrimary
-                        )
-                        Text(
-                            bookmark.dateAdded,
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun HighlightsList(highlights: List<HighlightInfo>) {
-    if (highlights.isEmpty()) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.XL),
-            contentAlignment = Alignment.Center
-        ) {
-            Text("No highlights yet", color = MediaColors.TextSecondary)
-        }
-    } else {
-        LazyColumn {
-            items(items = highlights, key = { it.id }) { highlight ->
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(MediaSpacing.MD)
-                ) {
-                    Surface(
-                        color = highlight.color.copy(alpha = 0.3f),
-                        shape = RoundedCornerShape(MediaCorners.XS)
-                    ) {
-                        Text(
-                            highlight.text,
-                            style = MediaTypography.BodySmall,
-                            color = MediaColors.TextPrimary,
-                            modifier = Modifier.padding(MediaSpacing.SM)
-                        )
-                    }
-                    Text(
-                        "Page ${highlight.pageNumber + 1}",
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextTertiary,
-                        modifier = Modifier.padding(top = MediaSpacing.XS)
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun AnnotationsList(annotations: List<AnnotationInfo>) {
-    if (annotations.isEmpty()) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.XL),
-            contentAlignment = Alignment.Center
-        ) {
-            Text("No notes yet", color = MediaColors.TextSecondary)
-        }
-    } else {
-        LazyColumn {
-            items(items = annotations, key = { it.id }) { annotation ->
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(MediaSpacing.MD)
-                ) {
-                    Text(
-                        annotation.note,
-                        style = MediaTypography.BodyMedium,
-                        color = MediaColors.TextPrimary
-                    )
-                    Row(
-                        modifier = Modifier.padding(top = MediaSpacing.XS),
-                        horizontalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-                    ) {
-                        Text(
-                            "Page ${annotation.pageNumber + 1}",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                        Text(
-                            annotation.dateAdded,
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // HELPER FUNCTIONS
@@ -1333,42 +519,6 @@ data class ReaderSearchResult(
     val startOffset: Int
 )
 
-data class ReaderSettings(
-    val fontSize: Float = 18f,
-    val fontFamily: ReaderFontFamily = ReaderFontFamily.SYSTEM,
-    val lineHeight: Float = 1.5f,
-    val marginHorizontal: Float = 16f,
-    val marginVertical: Float = 24f,
-    val textAlign: ReaderTextAlign = ReaderTextAlign.JUSTIFIED,
-    val themePreset: ReaderThemePreset = ReaderThemePreset.DARK,
-    val theme: ReaderTheme = ReaderThemePreset.DARK.toTheme(),
-    val readingMode: ReadingMode = ReadingMode.PAGINATED,
-    val pageAnimation: PageAnimation = PageAnimation.SLIDE,
-    val showPageNumber: Boolean = true,
-    val keepScreenOn: Boolean = true,
-    val volumePageTurn: Boolean = false,
-    val brightness: Float = 1f,
-    val warmth: Float = 0f
-)
-
-data class ReaderTheme(
-    val backgroundColor: Color,
-    val textColor: Color,
-    val linkColor: Color
-)
-
-enum class ReaderFontFamily(val displayName: String) {
-    SYSTEM("System"),
-    SERIF("Serif"),
-    SANS_SERIF("Sans Serif"),
-    MONOSPACE("Mono"),
-    GEORGIA("Georgia"),
-    TIMES("Times"),
-    PALATINO("Palatino"),
-    BOOKERLY("Bookerly"),
-    LITERATA("Literata"),
-    OPEN_DYSLEXIC("OpenDyslexic")
-}
 
 enum class ReaderTextAlign {
     LEFT, CENTER, RIGHT, JUSTIFIED
@@ -1387,34 +537,3 @@ enum class PageAnimation(val displayName: String) {
     FADE("Fade")
 }
 
-enum class ReaderThemePreset {
-    WHITE, SEPIA, DARK, AMOLED, NIGHT;
-    
-    fun toTheme(): ReaderTheme = when (this) {
-        WHITE -> ReaderTheme(
-            backgroundColor = Color.White,
-            textColor = Color.Black,
-            linkColor = Color(0xFF1976D2)
-        )
-        SEPIA -> ReaderTheme(
-            backgroundColor = Color(0xFFF4ECD8),
-            textColor = Color(0xFF5B4636),
-            linkColor = Color(0xFF8B5A2B)
-        )
-        DARK -> ReaderTheme(
-            backgroundColor = Color(0xFF1F1F1F),
-            textColor = Color(0xFFE0E0E0),
-            linkColor = Color(0xFF66B2FF)
-        )
-        AMOLED -> ReaderTheme(
-            backgroundColor = Color.Black,
-            textColor = Color(0xFFCCCCCC),
-            linkColor = Color(0xFF66B2FF)
-        )
-        NIGHT -> ReaderTheme(
-            backgroundColor = Color(0xFF1A0A0A),
-            textColor = Color(0xFFFFCCCC),
-            linkColor = Color(0xFFFF9999)
-        )
-    }
-}
