+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/SearchInBookSheet.kt
@@ -1,0 +1,90 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun SearchInBookSheet(
+    searchResults: List<ReaderSearchResult>,
+    onSearch: (String) -> Unit,
+    onResultClick: (ReaderSearchResult) -> Unit,
+    onDismiss: () -> Unit
+) {
+    var query by remember { mutableStateOf("") }
+    
+    ModalBottomSheet(
+        onDismissRequest = onDismiss,
+        containerColor = MediaColors.BackgroundElevated
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(MediaSpacing.MD)
+        ) {
+            OutlinedTextField(
+                value = query,
+                onValueChange = {
+                    query = it
+                    onSearch(it)
+                },
+                placeholder = { Text("Search in book...", color = MediaColors.TextTertiary) },
+                leadingIcon = { Icon(Icons.Default.Search, null, tint = MediaColors.TextSecondary) },
+                modifier = Modifier.fillMaxWidth(),
+                colors = OutlinedTextFieldDefaults.colors(
+                    focusedBorderColor = MediaColors.AccentPrimary,
+                    unfocusedBorderColor = MediaColors.Border
+                )
+            )
+            
+            Spacer(Modifier.height(MediaSpacing.MD))
+            
+            if (searchResults.isEmpty() && query.isNotEmpty()) {
+                Text(
+                    "No results found",
+                    style = MediaTypography.BodyMedium,
+                    color = MediaColors.TextSecondary,
+                    modifier = Modifier.padding(MediaSpacing.MD)
+                )
+            } else {
+                LazyColumn {
+                    items(items = searchResults, key = { "${it.pageNumber}-${it.excerpt.hashCode()}" }) { result ->
+                        SearchResultItem(
+                            result = result,
+                            onClick = { onResultClick(result) }
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
