+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/TwoPageContent.kt
@@ -1,0 +1,102 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+internal fun TwoPageContent(
+    pagerState: PagerState,
+    pages: List<BookPage>,
+    settings: ReaderSettings
+) {
+    val configuration = LocalConfiguration.current
+    val isLandscape = configuration.screenWidthDp > configuration.screenHeightDp
+    
+    if (isLandscape) {
+        HorizontalPager(
+            state = pagerState,
+            modifier = Modifier.fillMaxSize()
+        ) { spreadIndex ->
+            Row(modifier = Modifier.fillMaxSize()) {
+                // Left page
+                val leftPageIndex = spreadIndex * 2
+                Box(
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxHeight()
+                        .padding(start = settings.marginHorizontal.dp, end = 8.dp)
+                ) {
+                    pages.getOrNull(leftPageIndex)?.let { page ->
+                        BookPageContent(
+                            page = page,
+                            settings = settings,
+                            highlights = emptyList()
+                        )
+                    }
+                }
+                
+                // Divider
+                VerticalDivider(
+                    color = settings.theme.textColor.copy(alpha = 0.1f),
+                    modifier = Modifier.fillMaxHeight()
+                )
+                
+                // Right page
+                val rightPageIndex = leftPageIndex + 1
+                Box(
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxHeight()
+                        .padding(start = 8.dp, end = settings.marginHorizontal.dp)
+                ) {
+                    pages.getOrNull(rightPageIndex)?.let { page ->
+                        BookPageContent(
+                            page = page,
+                            settings = settings,
+                            highlights = emptyList()
+                        )
+                    }
+                }
+            }
+        }
+    } else {
+        // Fall back to single page in portrait
+        PaginatedContent(
+            pagerState = pagerState,
+            pages = pages,
+            settings = settings,
+            highlights = emptyList(),
+            pageAnimation = settings.pageAnimation
+        )
+    }
+}
