+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/reader/AdvancedComicReader.kt
@@ -296,187 +296,7 @@ fun AdvancedComicReader(
     }
 }
 
-@Composable
-private fun SinglePageView(
-    pages: List<String>,
-    currentPage: Int,
-    panels: List<Rect>,
-    currentPanel: Int,
-    panelByPanelMode: Boolean,
-    showPanelBorders: Boolean,
-    scale: Float,
-    offsetX: Float,
-    offsetY: Float,
-    onPageChanged: (Int) -> Unit,
-    onPanelSelected: (Int) -> Unit,
-    onScaleChanged: (Float) -> Unit,
-    onOffsetChanged: (Float, Float) -> Unit,
-    onTap: () -> Unit
-) {
-    if (currentPage > 0 && currentPage <= pages.size) {
-        val currentPageUri = pages[currentPage - 1]
 
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .graphicsLayer(
-                    scaleX = scale,
-                    scaleY = scale,
-                    translationX = offsetX,
-                    translationY = offsetY
-                )
-                .pointerInput(Unit) {
-                    detectTransformGestures { _, pan, zoom, _ ->
-                        if (!panelByPanelMode) {
-                            val newScale = (scale * zoom).coerceIn(0.5f, 5f)
-                            onScaleChanged(newScale)
-                            onOffsetChanged(offsetX + pan.x, offsetY + pan.y)
-                        }
-                    }
-                }
-                .pointerInput(Unit) {
-                    if (panelByPanelMode) {
-                        detectTapGestures { offset ->
-                            // In panel mode, find which panel was tapped
-                            val tappedPanel = panels.indexOfFirst { panel ->
-                                panel.contains(offset)
-                            }
-                            if (tappedPanel >= 0) {
-                                onPanelSelected(tappedPanel)
-                            }
-                            onTap()
-                        }
-                    } else {
-                        detectTapGestures(
-                            onTap = { onTap() },
-                            onDoubleTap = { offset ->
-                                // Double tap to zoom or fit to panel
-                                if (panels.isNotEmpty()) {
-                                    val tappedPanel = panels.indexOfFirst { panel ->
-                                        panel.contains(offset)
-                                    }
-                                    if (tappedPanel >= 0) {
-                                        onPanelSelected(tappedPanel)
-                                        // Zoom to fit panel
-                                        val panel = panels[tappedPanel]
-                                        val zoomScale = min(
-                                            size.width / panel.width,
-                                            size.height / panel.height
-                                        ) * 0.9f
-                                        onScaleChanged(zoomScale)
-                                        onOffsetChanged(-panel.left * zoomScale, -panel.top * zoomScale)
-                                    }
-                                }
-                            }
-                        )
-                    }
-                }
-        ) {
-            AsyncImage(
-                    
-                model = ImageRequest.Builder(LocalContext.current)
-                    .data(currentPageUri)
-                    .crossfade(true)
-                    .build(),
-                contentDescription = "Comic page $currentPage",
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Fit
-            )
-
-            // Draw panel borders and highlights
-            if ((showPanelBorders || panelByPanelMode) && panels.isNotEmpty()) {
-                Canvas(modifier = Modifier.fillMaxSize()) {
-                    panels.forEachIndexed { index, panel ->
-                        val isCurrentPanel = panelByPanelMode && index == currentPanel
-
-                        drawRect(
-                            color = if (isCurrentPanel) Color.Red else Color.Yellow,
-                            topLeft = Offset(panel.left, panel.top),
-                            size = Size(panel.width, panel.height),
-                            style = Stroke(
-                                width = if (isCurrentPanel) 4.dp.toPx() else 2.dp.toPx()
-                            )
-                        )
-
-                        if (showPanelBorders) {
-                            drawRect(
-                                color = Color.Black.copy(alpha = 0.1f),
-                                topLeft = Offset(panel.left, panel.top),
-                                size = Size(panel.width, panel.height)
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DoublePageView(
-    pages: List<String>,
-    currentPage: Int,
-    scale: Float,
-    offsetX: Float,
-    offsetY: Float,
-    onPageChanged: (Int) -> Unit,
-    onScaleChanged: (Float) -> Unit,
-    onOffsetChanged: (Float, Float) -> Unit,
-    onTap: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxSize()
-            .graphicsLayer(
-                scaleX = scale,
-                scaleY = scale,
-                translationX = offsetX,
-                translationY = offsetY
-            )
-            .pointerInput(Unit) {
-                detectTransformGestures { _, pan, zoom, _ ->
-                    val newScale = (scale * zoom).coerceIn(0.5f, 5f)
-                    onScaleChanged(newScale)
-                    onOffsetChanged(offsetX + pan.x, offsetY + pan.y)
-                }
-            }
-            .pointerInput(Unit) {
-                detectTapGestures(onTap = { onTap() })
-            }
-    ) {
-        // Left page
-        if (currentPage > 1) {
-            AsyncImage(
-                    
-                model = ImageRequest.Builder(LocalContext.current)
-                    .data(pages[currentPage - 2])
-                    .crossfade(true)
-                    .build(),
-                contentDescription = "Comic page ${currentPage - 1}",
-                modifier = Modifier.weight(1f),
-                contentScale = ContentScale.Fit
-            )
-        } else {
-            Spacer(modifier = Modifier.weight(1f))
-        }
-
-        // Right page
-        if (currentPage <= pages.size) {
-            AsyncImage(
-                    
-                model = ImageRequest.Builder(LocalContext.current)
-                    .data(pages[currentPage - 1])
-                    .crossfade(true)
-                    .build(),
-                contentDescription = "Comic page $currentPage",
-                modifier = Modifier.weight(1f),
-                contentScale = ContentScale.Fit
-            )
-        } else {
-            Spacer(modifier = Modifier.weight(1f))
-        }
-    }
-}
 
 @Composable
 private fun WebtoonView(
@@ -509,66 +329,6 @@ private fun WebtoonView(
     }
 }
 
-@Composable
-private fun PanelNavigationHints(
-    currentPanel: Int,
-    totalPanels: Int,
-    canGoNext: Boolean
-) {
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        // Next panel hint
-        if (canGoNext) {
-            AnimatedVisibility(
-                visible = true,
-                enter = fadeIn() + scaleIn(),
-                exit = fadeOut() + scaleOut(),
-                modifier = Modifier
-                    .align(Alignment.CenterEnd)
-                    .padding(16.dp)
-            ) {
-                Card(
-                    colors = CardDefaults.cardColors(
-                        containerColor = Color.Black.copy(alpha = 0.7f)
-                    ),
-                    shape = RoundedCornerShape(20.dp)
-                ) {
-                    Icon(
-                        Icons.AutoMirrored.Filled.ArrowForward,
-                        contentDescription = "Next panel",
-                        tint = Color.White,
-                        modifier = Modifier
-                            .padding(12.dp)
-                            .size(24.dp)
-                    )
-                }
-            }
-        }
-
-        // Panel indicator
-        AnimatedVisibility(
-            visible = totalPanels > 1,
-            modifier = Modifier
-                .align(Alignment.TopCenter)
-                .padding(16.dp)
-        ) {
-            Card(
-                colors = CardDefaults.cardColors(
-                    containerColor = Color.Black.copy(alpha = 0.7f)
-                ),
-                shape = RoundedCornerShape(16.dp)
-            ) {
-                Text(
-                    text = "${currentPanel + 1} / $totalPanels",
-                    color = Color.White,
-                    fontSize = 12.sp,
-                    modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp)
-                )
-            }
-        }
-    }
-}
 
 @Composable
 private fun LoadingState() {
@@ -693,260 +453,8 @@ private fun TopComicControls(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun BottomComicControls(
-    currentPage: Int,
-    totalPages: Int,
-    currentPanel: Int,
-    totalPanels: Int,
-    panelByPanelMode: Boolean,
-    readingDirection: ReadingDirection,
-    canGoPrevious: Boolean,
-    canGoNext: Boolean,
-    onPreviousPage: () -> Unit,
-    onNextPage: () -> Unit,
-    onPreviousPanel: () -> Unit,
-    onNextPanel: () -> Unit,
-    onPageSeek: (Int) -> Unit,
-    onTogglePanelMode: () -> Unit
-) {
-    Surface(
-        color = Color.Black.copy(alpha = 0.7f),
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Column(modifier = Modifier.padding(8.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                IconButton(onClick = if (panelByPanelMode) onPreviousPanel else onPreviousPage, enabled = canGoPrevious) {
-                    Icon(
-                        if (readingDirection == ReadingDirection.RIGHT_TO_LEFT) Icons.AutoMirrored.Filled.ArrowForward else Icons.AutoMirrored.Filled.ArrowBack,
-                        contentDescription = "Previous",
-                        tint = if (canGoPrevious) Color.White else Color.Gray
-                    )
-                }
-                Text("$currentPage / $totalPages", color = Color.White)
-                IconButton(onClick = if (panelByPanelMode) onNextPanel else onNextPage, enabled = canGoNext) {
-                    Icon(
-                        if (readingDirection == ReadingDirection.RIGHT_TO_LEFT) Icons.AutoMirrored.Filled.ArrowBack else Icons.AutoMirrored.Filled.ArrowForward,
-                        contentDescription = "Next",
-                        tint = if (canGoNext) Color.White else Color.Gray
-                    )
-                }
-                IconButton(onClick = onTogglePanelMode) {
-                    Icon(
-                        if (panelByPanelMode) Icons.Default.ViewAgenda else Icons.Default.Dashboard,
-                        contentDescription = "Toggle Panel Mode",
-                        tint = Color.White
-                    )
-                }
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ComicReaderSettingsSheet(
-    settings: ViewerSettings,
-    comicSettings: ComicSettings,
-    onDismiss: () -> Unit,
-    onSettingsChanged: (ViewerSettings) -> Unit,
-    onComicSettingsChanged: (ComicSettings) -> Unit
-) {
-    ModalBottomSheet(onDismissRequest = onDismiss) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            Text("Comic Reader Settings", style = MaterialTheme.typography.headlineSmall)
-            Spacer(modifier = Modifier.height(16.dp))
 
-            Text("Reading Mode", style = MaterialTheme.typography.titleMedium)
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.fillMaxWidth()) {
-                ReadingMode.values().forEach { mode ->
-                    FilterChip(
-                        selected = comicSettings.readingMode == mode,
-                        onClick = { onComicSettingsChanged(comicSettings.copy(readingMode = mode)) },
-                        label = { Text(mode.name.replace("_", " ")) }
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text("Reading Direction", style = MaterialTheme.typography.titleMedium)
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.fillMaxWidth()) {
-                ReadingDirection.values().forEach { direction ->
-                    FilterChip(
-                        selected = comicSettings.readingDirection == direction,
-                        onClick = { onComicSettingsChanged(comicSettings.copy(readingDirection = direction)) },
-                        label = { Text(direction.name.replace("_", " ")) }
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text("Fit Mode", style = MaterialTheme.typography.titleMedium)
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.fillMaxWidth()) {
-                FitMode.values().forEach { mode ->
-                    FilterChip(
-                        selected = comicSettings.fitMode == mode,
-                        onClick = { onComicSettingsChanged(comicSettings.copy(fitMode = mode)) },
-                        label = { Text(mode.name.replace("_", " ")) }
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text("Zoom Limit: ${String.format("%.1f", comicSettings.maxZoom)}x", style = MaterialTheme.typography.titleMedium)
-            Slider(
-                value = comicSettings.maxZoom,
-                onValueChange = { onComicSettingsChanged(comicSettings.copy(maxZoom = it)) },
-                valueRange = 1.5f..6f,
-                steps = 9
-            )
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text("Reader Comfort", style = MaterialTheme.typography.titleMedium)
-            SettingToggleRow(
-                title = "Night Mode",
-                checked = settings.nightMode,
-                onCheckedChange = { onSettingsChanged(settings.copy(nightMode = it)) }
-            )
-            SettingToggleRow(
-                title = "Blue Light Filter",
-                checked = settings.blueLightFilter,
-                onCheckedChange = { onSettingsChanged(settings.copy(blueLightFilter = it)) }
-            )
-            SettingToggleRow(
-                title = "Keep Screen On",
-                checked = settings.keepScreenOn,
-                onCheckedChange = { onSettingsChanged(settings.copy(keepScreenOn = it)) }
-            )
-
-            Spacer(modifier = Modifier.height(12.dp))
-            Text("Panel Navigation", style = MaterialTheme.typography.titleMedium)
-            SettingToggleRow(
-                title = "Panel Detection",
-                checked = comicSettings.panelDetection,
-                onCheckedChange = { onComicSettingsChanged(comicSettings.copy(panelDetection = it)) }
-            )
-            SettingToggleRow(
-                title = "Panel-by-Panel Mode",
-                checked = comicSettings.panelByPanelMode,
-                onCheckedChange = { onComicSettingsChanged(comicSettings.copy(panelByPanelMode = it)) }
-            )
-            SettingToggleRow(
-                title = "Show Panel Borders",
-                checked = comicSettings.showPanelBorders,
-                onCheckedChange = { onComicSettingsChanged(comicSettings.copy(showPanelBorders = it)) }
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-            Button(onClick = onDismiss, modifier = Modifier.fillMaxWidth()) {
-                Text("Done")
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun PanelBrowserSheet(
-    pages: List<String>,
-    currentPage: Int,
-    panels: Map<Int, List<Rect>>,
-    onDismiss: () -> Unit,
-    onPageSelected: (Int) -> Unit,
-    onPanelSelected: (Int, Int) -> Unit
-) {
-    ModalBottomSheet(onDismissRequest = onDismiss) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            Text("Panel Browser", style = MaterialTheme.typography.headlineSmall)
-            Spacer(modifier = Modifier.height(16.dp))
-
-            LazyVerticalGrid(
-                columns = GridCells.Fixed(3),
-                contentPadding = PaddingValues(8.dp),
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-                modifier = Modifier.fillMaxHeight()
-            ) {
-                items(pages.size) { index ->
-                    val pageNumber = index + 1
-                    val panelCount = panels[pageNumber]?.size ?: 0
-
-                    Card(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onPageSelected(pageNumber) },
-                        colors = CardDefaults.cardColors(
-                            containerColor = if (pageNumber == currentPage) {
-                                MaterialTheme.colorScheme.primaryContainer
-                            } else {
-                                MaterialTheme.colorScheme.surface
-                            }
-                        )
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(8.dp),
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            AsyncImage(
-                                model = pages[index],
-                                contentDescription = "Page $pageNumber",
-                                modifier = Modifier
-                                    .height(120.dp)
-                                    .fillMaxWidth()
-                                    .clip(RoundedCornerShape(8.dp))
-                            )
-                            Spacer(modifier = Modifier.height(8.dp))
-                            Text(
-                                "Page $pageNumber",
-                                style = MaterialTheme.typography.bodyMedium,
-                                fontWeight = FontWeight.Medium
-                            )
-                            if (panelCount > 0) {
-                                Text(
-                                    "$panelCount panels",
-                                    style = MaterialTheme.typography.labelSmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                                Row(
-                                    horizontalArrangement = Arrangement.spacedBy(4.dp),
-                                    modifier = Modifier.padding(top = 4.dp)
-                                ) {
-                                    panels[pageNumber]?.take(3)?.forEachIndexed { panelIndex, _ ->
-                                        AssistChip(
-                                            onClick = { onPanelSelected(pageNumber, panelIndex) },
-                                            label = { Text("#${panelIndex + 1}") }
-                                        )
-                                    }
-                                    if (panelCount > 3) {
-                                        AssistChip(
-                                            onClick = { onPageSelected(pageNumber) },
-                                            label = { Text("+${panelCount - 3}") }
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun SettingToggleRow(
