+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/reader/EnhancedEReaderScreen.kt
@@ -488,69 +488,6 @@ fun EnhancedEReaderScreen(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun AIInsightsSheet(
-    insights: List<ReaderAIInsight>,
-    onDismiss: () -> Unit
-) {
-    ModalBottomSheet(onDismissRequest = onDismiss) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(24.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "AI Insights",
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold
-                )
-                Icon(Icons.Default.AutoAwesome, contentDescription = "AI", tint = MaterialTheme.colorScheme.primary)
-            }
-
-            if (insights.isEmpty()) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(200.dp),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = "No insights available yet.\nContinue reading to generate insights.",
-                        textAlign = TextAlign.Center,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-            } else {
-                LazyColumn(
-                    verticalArrangement = Arrangement.spacedBy(16.dp)
-                ) {
-                    items(insights) { insight ->
-                        insight.summary?.let {
-                            InsightCard("Summary", it)
-                        }
-                        insight.characterAnalysis?.let {
-                            InsightCard("Character Analysis", it)
-                        }
-                        insight.perspectiveAnalysis?.let {
-                            InsightCard("Perspective Analysis", it)
-                        }
-                        if (insight.keyThemes.isNotEmpty()) {
-                            InsightCard("Key Themes", insight.keyThemes.joinToString("\n• ", prefix = "• "))
-                        }
-                    }
-                }
-            }
-            Spacer(modifier = Modifier.height(24.dp))
-        }
-    }
-}
 
 @Composable
 private fun InsightCard(title: String, content: String) {
@@ -618,175 +555,7 @@ private fun LoadingView() {
     }
 }
 
-@Composable
-private fun ErrorView(message: String, onRetry: () -> Unit) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-            modifier = Modifier.padding(32.dp)
-        ) {
-            Icon(
-                Icons.Default.Error,
-                contentDescription = "Media image",
-                modifier = Modifier.size(64.dp),
-                tint = MaterialTheme.colorScheme.error
-            )
-            Text(
-                text = message,
-                style = MaterialTheme.typography.bodyLarge,
-                color = MaterialTheme.colorScheme.error,
-                textAlign = TextAlign.Center
-            )
-            Button(onClick = onRetry) {
-                Icon(Icons.Default.Refresh, contentDescription = "Refresh")
-                Spacer(modifier = Modifier.width(8.dp))
-                Text("Retry")
-            }
-        }
-    }
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ReadingSettingsSheet(
-    fontSize: TextUnit,
-    onFontSizeChange: (TextUnit) -> Unit,
-    fontFamily: GenericFontFamily,
-    onFontFamilyChange: (GenericFontFamily) -> Unit,
-    backgroundColor: Color,
-    onBackgroundChange: (Color, Color) -> Unit,
-    brightness: Float,
-    onBrightnessChange: (Float) -> Unit,
-    onDismiss: () -> Unit
-) {
-    ModalBottomSheet(onDismissRequest = onDismiss) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(24.dp),
-            verticalArrangement = Arrangement.spacedBy(24.dp)
-        ) {
-            Text(
-                text = "Reading Settings",
-                style = MaterialTheme.typography.titleLarge,
-                fontWeight = FontWeight.Bold
-            )
-
-            // Font size
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text("Font Size", style = MaterialTheme.typography.titleMedium)
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    IconButton(onClick = { onFontSizeChange((fontSize.value - 2).sp) }) {
-                        Icon(Icons.Default.Remove, contentDescription = "Decrease")
-                    }
-                    Text(
-                        text = "${fontSize.value.toInt()} sp",
-                        modifier = Modifier.weight(1f),
-                        textAlign = TextAlign.Center
-                    )
-                    IconButton(onClick = { onFontSizeChange((fontSize.value + 2).sp) }) {
-                        Icon(Icons.Default.Add, contentDescription = "Increase")
-                    }
-                }
-            }
-
-            // Font family
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text("Font", style = MaterialTheme.typography.titleMedium)
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    FilterChip(
-                        selected = fontFamily == FontFamily.Serif,
-                        onClick = { onFontFamilyChange(FontFamily.Serif) },
-                        label = { Text("Serif") }
-                    )
-                    FilterChip(
-                        selected = fontFamily == FontFamily.SansSerif,
-                        onClick = { onFontFamilyChange(FontFamily.SansSerif) },
-                        label = { Text("Sans") }
-                    )
-                    FilterChip(
-                        selected = fontFamily == FontFamily.Monospace,
-                        onClick = { onFontFamilyChange(FontFamily.Monospace) },
-                        label = { Text("Mono") }
-                    )
-                }
-            }
-
-            // Theme
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text("Theme", style = MaterialTheme.typography.titleMedium)
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    ThemeButton(
-                        name = "Sepia",
-                        backgroundColor = Color(0xFFFFFBF0),
-                        textColor = Color(0xFF3E2723),
-                        isSelected = backgroundColor == Color(0xFFFFFBF0),
-                        onClick = { onBackgroundChange(Color(0xFFFFFBF0), Color(0xFF3E2723)) }
-                    )
-                    ThemeButton(
-                        name = "Night",
-                        backgroundColor = Color(0xFF1A1A1A),
-                        textColor = Color(0xFFE0E0E0),
-                        isSelected = backgroundColor == Color(0xFF1A1A1A),
-                        onClick = { onBackgroundChange(Color(0xFF1A1A1A), Color(0xFFE0E0E0)) }
-                    )
-                    ThemeButton(
-                        name = "White",
-                        backgroundColor = Color.White,
-                        textColor = Color.Black,
-                        isSelected = backgroundColor == Color.White,
-                        onClick = { onBackgroundChange(Color.White, Color.Black) }
-                    )
-                }
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    ThemeButton(
-                        name = "Paper",
-                        backgroundColor = Color(0xFFF5F5DC), // Beige/Cream
-                        textColor = Color(0xFF2C2C2C),
-                        isSelected = backgroundColor == Color(0xFFF5F5DC),
-                        onClick = { onBackgroundChange(Color(0xFFF5F5DC), Color(0xFF2C2C2C)) }
-                    )
-                    ThemeButton(
-                        name = "OLED",
-                        backgroundColor = Color.Black,
-                        textColor = Color(0xFFB0B0B0),
-                        isSelected = backgroundColor == Color.Black,
-                        onClick = { onBackgroundChange(Color.Black, Color(0xFFB0B0B0)) }
-                    )
-                }
-            }
-
-            // Brightness
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text("Brightness", style = MaterialTheme.typography.titleMedium)
-                Slider(
-                    value = brightness,
-                    onValueChange = onBrightnessChange,
-                    valueRange = 0.3f..1f
-                )
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-        }
-    }
-}
 
 @Composable
 private fun ThemeButton(
