+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/reader/EnhancedPDFReaderScreen.kt
@@ -664,68 +664,6 @@ fun EnhancedPDFReaderScreen(
     }
 }
 
-@Composable
-fun PDFTextContent(
-    pageContent: String,
-    pageNumber: Int,
-    totalPages: Int,
-    showTextSelectionMode: Boolean,
-    onTextSelected: (String) -> Unit
-) {
-    val scrollState = rememberScrollState()
-    
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        shape = RoundedCornerShape(8.dp),
-        color = Color.White,
-        shadowElevation = 4.dp
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .verticalScroll(scrollState)
-                .padding(24.dp)
-        ) {
-            // Page header
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "Page $pageNumber",
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold,
-                    color = Color.Black
-                )
-                Text(
-                    text = "$pageNumber / $totalPages",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = Color.Gray
-                )
-            }
-            
-            HorizontalDivider(
-                modifier = Modifier.padding(vertical = 16.dp),
-                color = Color.LightGray
-            )
-            
-            // Page content
-            SelectionContainer(enabled = showTextSelectionMode) {
-                Text(
-                    text = pageContent,
-                    style = MaterialTheme.typography.bodyLarge.copy(
-                        lineHeight = 28.sp,
-                        color = Color.Black
-                    ),
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-    }
-}
 
 @Composable
 fun SelectionContainer(
@@ -741,255 +679,8 @@ fun SelectionContainer(
     }
 }
 
-@Composable
-fun PDFSearchBar(
-    query: String,
-    onQueryChange: (String) -> Unit,
-    onSearch: () -> Unit,
-    onDismiss: () -> Unit,
-    searchResults: List<SearchResult>,
-    onResultClick: (SearchResult) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(8.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            OutlinedTextField(
-                value = query,
-                onValueChange = onQueryChange,
-                placeholder = { Text("Search in PDF...") },
-                leadingIcon = { Icon(Icons.Default.Search, contentDescription = "Search") },
-                trailingIcon = {
-                    Row {
-                        if (query.isNotEmpty()) {
-                            IconButton(onClick = { onQueryChange("") }) {
-                                Icon(Icons.Default.Clear, contentDescription = "Clear")
-                            }
-                        }
-                        IconButton(onClick = onDismiss) {
-                            Icon(Icons.Default.Close, contentDescription = "Close")
-                        }
-                    }
-                },
-                singleLine = true,
-                modifier = Modifier.fillMaxWidth(),
-                keyboardActions = androidx.compose.foundation.text.KeyboardActions(
-                    onSearch = { onSearch() }
-                )
-            )
 
-            if (searchResults.isNotEmpty()) {
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(
-                    "${searchResults.size} results found",
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.primary
-                )
 
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                searchResults.take(5).forEach { result ->
-                    Surface(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(vertical = 4.dp)
-                            .clickable { onResultClick(result) },
-                        shape = RoundedCornerShape(8.dp),
-                        color = MaterialTheme.colorScheme.surfaceVariant
-                    ) {
-                        Row(
-                            modifier = Modifier.padding(12.dp),
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Surface(
-                                shape = CircleShape,
-                                color = MaterialTheme.colorScheme.primary
-                            ) {
-                                Text(
-                                    text = "${result.pageNumber}",
-                                    style = MaterialTheme.typography.labelMedium,
-                                    color = MaterialTheme.colorScheme.onPrimary,
-                                    modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp)
-                                )
-                            }
-                            Spacer(modifier = Modifier.width(12.dp))
-                            Text(
-                                text = result.context,
-                                style = MaterialTheme.typography.bodySmall,
-                                maxLines = 2
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun PDFAnnotationOverlay(
-    annotations: List<Annotation>,
-    annotationTool: AnnotationTool,
-    onAnnotationCreated: (Annotation) -> Unit,
-    currentPage: Int
-) {
-    var drawingPath by remember { mutableStateOf<Path?>(null) }
-    var currentPath by remember { mutableStateOf(Path()) }
-
-    Canvas(
-        modifier = Modifier
-            .fillMaxSize()
-            .pointerInput(annotationTool) {
-                if (annotationTool == AnnotationTool.DRAW) {
-                    detectDragGestures(
-                        onDragStart = { offset ->
-                            currentPath = Path()
-                            currentPath.moveTo(offset.x, offset.y)
-                        },
-                        onDrag = { change, _ ->
-                            currentPath.lineTo(change.position.x, change.position.y)
-                            drawingPath = Path().apply { addPath(currentPath) }
-                        },
-                        onDragEnd = {
-                            onAnnotationCreated(
-                                Annotation(
-                                    id = System.currentTimeMillis(),
-                                    type = AnnotationType.DRAWING,
-                                    pageNumber = currentPage,
-                                    content = "",
-                                    position = Offset.Zero,
-                                    path = drawingPath
-                                )
-                            )
-                            drawingPath = null
-                        }
-                    )
-                }
-            }
-    ) {
-        // Draw existing annotations
-        annotations.forEach { annotation ->
-            when (annotation.type) {
-                AnnotationType.HIGHLIGHT -> {
-                    drawRect(
-                        color = Color.Yellow.copy(alpha = 0.4f),
-                        topLeft = annotation.position,
-                        size = annotation.size ?: androidx.compose.ui.geometry.Size(200f, 30f)
-                    )
-                }
-                AnnotationType.DRAWING -> {
-                    annotation.path?.let { path ->
-                        drawPath(
-                            path = path,
-                            color = Color.Red,
-                            style = Stroke(width = 4.dp.toPx())
-                        )
-                    }
-                }
-                AnnotationType.NOTE -> {
-                    drawCircle(
-                        color = Color.Blue,
-                        radius = 12.dp.toPx(),
-                        center = annotation.position
-                    )
-                }
-                AnnotationType.STRIKETHROUGH -> {
-                    drawLine(
-                        color = Color.Red,
-                        start = annotation.position,
-                        end = annotation.position.copy(x = annotation.position.x + 150f),
-                        strokeWidth = 3.dp.toPx()
-                    )
-                }
-            }
-        }
-
-        // Draw current drawing
-        drawingPath?.let { path ->
-            drawPath(
-                path = path,
-                color = Color.Red,
-                style = Stroke(width = 4.dp.toPx())
-            )
-        }
-    }
-}
-
-@Composable
-fun PageThumbnailStrip(
-    totalPages: Int,
-    currentPage: Int,
-    onPageSelect: (Int) -> Unit,
-    onDismiss: () -> Unit
-) {
-    val listState = rememberLazyListState()
-    val scope = rememberCoroutineScope()
-    
-    LaunchedEffect(currentPage) {
-        scope.launch {
-            listState.animateScrollToItem((currentPage - 1).coerceAtLeast(0))
-        }
-    }
-    
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        color = MaterialTheme.colorScheme.surface,
-        tonalElevation = 8.dp
-    ) {
-        Column(modifier = Modifier.padding(8.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "Page Thumbnails",
-                    style = MaterialTheme.typography.titleSmall
-                )
-                IconButton(onClick = onDismiss) {
-                    Icon(Icons.Default.Close, contentDescription = "Close")
-                }
-            }
-            
-            LazyRow(
-                state = listState,
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                contentPadding = PaddingValues(vertical = 8.dp)
-            ) {
-                items(totalPages) { index ->
-                    val pageNum = index + 1
-                    val isSelected = pageNum == currentPage
-                    
-                    Surface(
-                        modifier = Modifier
-                            .size(60.dp, 80.dp)
-                            .clickable { onPageSelect(pageNum) },
-                        shape = RoundedCornerShape(4.dp),
-                        color = if (isSelected) MaterialTheme.colorScheme.primaryContainer 
-                                else MaterialTheme.colorScheme.surfaceVariant,
-                        border = if (isSelected) 
-                            androidx.compose.foundation.BorderStroke(2.dp, MaterialTheme.colorScheme.primary)
-                        else null
-                    ) {
-                        Box(
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Text(
-                                text = "$pageNum",
-                                style = MaterialTheme.typography.titleMedium,
-                                fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Normal
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 fun TextSelectionActionBar(
@@ -1022,221 +713,8 @@ fun TextSelectionActionBar(
     }
 }
 
-@Composable
-fun GoToPageDialog(
-    currentPage: Int,
-    totalPages: Int,
-    onGoToPage: (Int) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var pageInput by remember { mutableStateOf(currentPage.toString()) }
-    
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Go to Page") },
-        text = {
-            Column {
-                Text("Enter page number (1-$totalPages):")
-                Spacer(modifier = Modifier.height(16.dp))
-                OutlinedTextField(
-                    value = pageInput,
-                    onValueChange = { 
-                        if (it.all { char -> char.isDigit() }) {
-                            pageInput = it
-                        }
-                    },
-                    label = { Text("Page") },
-                    singleLine = true,
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    val page = pageInput.toIntOrNull()
-                    if (page != null && page in 1..totalPages) {
-                        onGoToPage(page)
-                    }
-                }
-            ) {
-                Text("Go")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun PDFReaderSettingsSheet(
-    currentZoom: Float,
-    zoomMode: ZoomMode,
-    onZoomChange: (Float) -> Unit,
-    onZoomModeChange: (ZoomMode) -> Unit,
-    onDismiss: () -> Unit
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismiss
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(24.dp)
-        ) {
-            Text(
-                text = "Reader Settings",
-                style = MaterialTheme.typography.headlineSmall,
-                fontWeight = FontWeight.Bold
-            )
-            
-            Spacer(modifier = Modifier.height(24.dp))
-            
-            // Zoom level
-            Text(
-                text = "Zoom Level: ${(currentZoom * 100).toInt()}%",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Slider(
-                value = currentZoom,
-                onValueChange = onZoomChange,
-                valueRange = 0.5f..5f,
-                steps = 17,
-                modifier = Modifier.padding(vertical = 8.dp)
-            )
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween
-            ) {
-                Text("50%", style = MaterialTheme.typography.bodySmall)
-                Text("500%", style = MaterialTheme.typography.bodySmall)
-            }
-            
-            Spacer(modifier = Modifier.height(24.dp))
-            
-            // Zoom mode
-            Text(
-                text = "Zoom Mode",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                ZoomMode.entries.forEach { mode ->
-                    FilterChip(
-                        selected = zoomMode == mode,
-                        onClick = { onZoomModeChange(mode) },
-                        label = { 
-                            Text(
-                                when (mode) {
-                                    ZoomMode.FIT_WIDTH -> "Fit Width"
-                                    ZoomMode.FIT_PAGE -> "Fit Page"
-                                    ZoomMode.CUSTOM -> "Custom"
-                                }
-                            )
-                        }
-                    )
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(32.dp))
-        }
-    }
-}
 
-@Composable
-fun FloatingAnnotationPanel(
-    annotations: List<Annotation>,
-    onAnnotationClick: (Annotation) -> Unit,
-    onAnnotationDelete: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier.width(220.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 8.dp)
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    "Annotations",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Medium
-                )
-                Text(
-                    "${annotations.size}",
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }
-
-            Spacer(modifier = Modifier.height(12.dp))
-
-            if (annotations.isEmpty()) {
-                Text(
-                    "No annotations on this page",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            } else {
-                annotations.forEach { annotation ->
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onAnnotationClick(annotation) }
-                            .padding(vertical = 6.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            when (annotation.type) {
-                                AnnotationType.HIGHLIGHT -> Icons.Default.FormatColorFill
-                                AnnotationType.NOTE -> Icons.AutoMirrored.Filled.Note
-                                AnnotationType.DRAWING -> Icons.Default.Draw
-                                AnnotationType.STRIKETHROUGH -> Icons.Default.FormatStrikethrough
-                            },
-                            contentDescription = null,
-                            modifier = Modifier.size(18.dp),
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-
-                        Spacer(modifier = Modifier.width(8.dp))
-
-                        Text(
-                            text = annotation.content.ifEmpty { annotation.type.name.lowercase().replaceFirstChar { it.uppercase() } },
-                            style = MaterialTheme.typography.bodySmall,
-                            modifier = Modifier.weight(1f),
-                            maxLines = 1
-                        )
-
-                        IconButton(
-                            onClick = { onAnnotationDelete(annotation.id) },
-                            modifier = Modifier.size(28.dp)
-                        ) {
-                            Icon(
-                                Icons.Default.Delete,
-                                contentDescription = "Delete",
-                                modifier = Modifier.size(18.dp),
-                                tint = MaterialTheme.colorScheme.error
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 // Data classes and enums for PDF reader
 enum class AnnotationTool(val displayName: String, val icon: androidx.compose.ui.graphics.vector.ImageVector) {
