+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/reader/components/DocumentReaderComponents.kt
@@ -166,504 +166,7 @@ fun TopReaderControls(
     }
 }
 
-@Composable
-fun BottomReaderControls(
-    currentPage: Int,
-    totalPages: Int,
-    readingProgress: Float,
-    canGoPrevious: Boolean,
-    canGoNext: Boolean,
-    onPreviousPage: () -> Unit,
-    onNextPage: () -> Unit,
-    onPageSeek: (Int) -> Unit
-) {
-    val previousAlpha = if (canGoPrevious) ENABLED_ALPHA else DISABLED_ALPHA
-    val nextAlpha = if (canGoNext) ENABLED_ALPHA else DISABLED_ALPHA
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(100.dp)
-            .background(
-                Brush.verticalGradient(
-                    colors = listOf(
-                        Color.Transparent,
-                        Color.Black.copy(alpha = 0.7f)
-                    )
-                )
-            )
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.BottomStart)
-                .padding(16.dp)
-        ) {
-            // Progress slider
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = currentPage.toString(),
-                    color = Color.White,
-                    style = MaterialTheme.typography.bodySmall,
-                    modifier = Modifier.width(40.dp),
-                    textAlign = TextAlign.Center
-                )
 
-                Slider(
-                    value = currentPage.toFloat(),
-                    onValueChange = { onPageSeek(it.toInt()) },
-                    valueRange = 1f..totalPages.toFloat(),
-                    modifier = Modifier.weight(1f),
-                    colors = SliderDefaults.colors(
-                        thumbColor = MaterialTheme.colorScheme.primary,
-                        activeTrackColor = MaterialTheme.colorScheme.primary,
-                        inactiveTrackColor = Color.White.copy(alpha = 0.3f)
-                    )
-                )
-
-                Text(
-                    text = totalPages.toString(),
-                    color = Color.White,
-                    style = MaterialTheme.typography.bodySmall,
-                    modifier = Modifier.width(40.dp),
-                    textAlign = TextAlign.Center
-                )
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // Navigation buttons
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    modifier = Modifier.widthIn(min = 72.dp)
-                ) {
-                    IconButton(
-                        onClick = onPreviousPage,
-                        enabled = canGoPrevious,
-                        modifier = Modifier.size(56.dp)
-                    ) {
-                        Icon(
-                            Icons.AutoMirrored.Filled.NavigateBefore,
-                            contentDescription = "Previous Page",
-                            tint = Color.White.copy(alpha = previousAlpha),
-                            modifier = Modifier.size(32.dp)
-                        )
-                    }
-                    Text(
-                        text = "Previous",
-                        style = MaterialTheme.typography.labelSmall,
-                        color = Color.White.copy(alpha = previousAlpha)
-                    )
-                }
-
-                // Progress indicator
-                LinearProgressIndicator(
-                    progress = readingProgress,
-                    modifier = Modifier
-                        .weight(1f)
-                        .height(4.dp)
-                        .clip(RoundedCornerShape(2.dp)),
-                    color = MaterialTheme.colorScheme.primary,
-                    trackColor = Color.White.copy(alpha = 0.3f)
-                )
-
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    modifier = Modifier.widthIn(min = 72.dp)
-                ) {
-                    IconButton(
-                        onClick = onNextPage,
-                        enabled = canGoNext,
-                        modifier = Modifier.size(56.dp)
-                    ) {
-                        Icon(
-                            Icons.AutoMirrored.Filled.NavigateNext,
-                            contentDescription = "Next Page",
-                            tint = Color.White.copy(alpha = nextAlpha),
-                            modifier = Modifier.size(32.dp)
-                        )
-                    }
-                    Text(
-                        text = "Next",
-                        style = MaterialTheme.typography.labelSmall,
-                        color = Color.White.copy(alpha = nextAlpha)
-                    )
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DocumentReaderSettingsSheet(
-    settings: ViewerSettings,
-    documentType: MediaViewerManager.DocumentFormat,
-    onDismiss: () -> Unit,
-    onSettingsChanged: (ViewerSettings) -> Unit
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        modifier = Modifier.fillMaxHeight(0.8f)
-    ) {
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            item {
-                Text(
-                    text = "Reading Settings",
-                    style = MaterialTheme.typography.headlineSmall,
-                    fontWeight = FontWeight.Medium
-                )
-            }
-
-            // Display Settings
-            item {
-                SettingsCard(title = "Display") {
-                    Column(
-                        verticalArrangement = Arrangement.spacedBy(16.dp)
-                    ) {
-                        // Brightness
-                        Text("Brightness: ${(settings.brightness * 100).toInt()}%")
-                        Slider(
-                            value = settings.brightness,
-                            onValueChange = {
-                                onSettingsChanged(settings.copy(brightness = it))
-                            },
-                            valueRange = 0.1f..1f
-                        )
-
-                        // Night mode
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Night Mode")
-                            Switch(
-                                checked = settings.nightMode,
-                                onCheckedChange = {
-                                    val newSettings = if (it) {
-                                        settings.copy(
-                                            nightMode = true,
-                                            backgroundColor = Color.Black,
-                                            textColor = Color.White
-                                        )
-                                    } else {
-                                        settings.copy(
-                                            nightMode = false,
-                                            backgroundColor = Color.White,
-                                            textColor = Color.Black
-                                        )
-                                    }
-                                    onSettingsChanged(newSettings)
-                                }
-                            )
-                        }
-
-                        // Blue light filter
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Blue Light Filter")
-                            Switch(
-                                checked = settings.blueLightFilter,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(blueLightFilter = it))
-                                }
-                            )
-                        }
-
-                        // Keep screen on
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Keep Screen On")
-                            Switch(
-                                checked = settings.keepScreenOn,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(keepScreenOn = it))
-                                }
-                            )
-                        }
-                    }
-                }
-            }
-
-            // Typography Settings (for text-based documents)
-            if (documentType != MediaViewerManager.DocumentFormat.PDF) {
-                item {
-                    SettingsCard(title = "Typography") {
-                        Column(
-                            verticalArrangement = Arrangement.spacedBy(16.dp)
-                        ) {
-                            // Font size
-                            Text("Font Size: ${settings.fontSize.toInt()}sp")
-                            Slider(
-                                value = settings.fontSize,
-                                onValueChange = {
-                                    onSettingsChanged(settings.copy(fontSize = it))
-                                },
-                                valueRange = 10f..36f,
-                                steps = 25
-                            )
-
-                            // Line spacing
-                            Text("Line Spacing: ${(settings.lineSpacing * 100).toInt()}%")
-                            Slider(
-                                value = settings.lineSpacing,
-                                onValueChange = {
-                                    onSettingsChanged(settings.copy(lineSpacing = it))
-                                },
-                                valueRange = 1.0f..3.0f,
-                                steps = 19
-                            )
-
-                            // Font family
-                            Text("Font Family")
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.spacedBy(8.dp)
-                            ) {
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(fontFamily = FontFamily.Default))
-                                    },
-                                    label = { Text("Default") },
-                                    selected = settings.fontFamily == FontFamily.Default
-                                )
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(fontFamily = FontFamily.Serif))
-                                    },
-                                    label = { Text("Serif") },
-                                    selected = settings.fontFamily == FontFamily.Serif
-                                )
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(fontFamily = FontFamily.SansSerif))
-                                    },
-                                    label = { Text("Sans Serif") },
-                                    selected = settings.fontFamily == FontFamily.SansSerif
-                                )
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(fontFamily = FontFamily.Monospace))
-                                    },
-                                    label = { Text("Mono") },
-                                    selected = settings.fontFamily == FontFamily.Monospace
-                                )
-                            }
-
-                            // Text alignment
-                            Text("Text Alignment")
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.spacedBy(8.dp)
-                            ) {
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(textAlignment = TextAlignment.LEFT))
-                                    },
-                                    label = { Text("Left") },
-                                    selected = settings.textAlignment == TextAlignment.LEFT
-                                )
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(textAlignment = TextAlignment.CENTER))
-                                    },
-                                    label = { Text("Center") },
-                                    selected = settings.textAlignment == TextAlignment.CENTER
-                                )
-                                FilterChip(
-                                    onClick = {
-                                        onSettingsChanged(settings.copy(textAlignment = TextAlignment.JUSTIFY))
-                                    },
-                                    label = { Text("Justify") },
-                                    selected = settings.textAlignment == TextAlignment.JUSTIFY
-                                )
-                            }
-
-                            // Margins
-                            Text("Horizontal Margin: ${settings.marginHorizontal.toInt()}dp")
-                            Slider(
-                                value = settings.marginHorizontal,
-                                onValueChange = {
-                                    onSettingsChanged(settings.copy(marginHorizontal = it))
-                                },
-                                valueRange = 8f..48f,
-                                steps = 39
-                            )
-                        }
-                    }
-                }
-            }
-
-            // Reading Settings
-            item {
-                SettingsCard(title = "Reading") {
-                    Column(
-                        verticalArrangement = Arrangement.spacedBy(16.dp)
-                    ) {
-                        // Reading mode
-                        Text("Reading Mode")
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingMode = ReadingMode.PAGE_BY_PAGE))
-                                },
-                                label = { Text("Page") },
-                                selected = settings.readingMode == ReadingMode.PAGE_BY_PAGE
-                            )
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingMode = ReadingMode.CONTINUOUS_VERTICAL))
-                                },
-                                label = { Text("Vertical") },
-                                selected = settings.readingMode == ReadingMode.CONTINUOUS_VERTICAL
-                            )
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingMode = ReadingMode.CONTINUOUS_HORIZONTAL))
-                                },
-                                label = { Text("Horizontal") },
-                                selected = settings.readingMode == ReadingMode.CONTINUOUS_HORIZONTAL
-                            )
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingMode = ReadingMode.WEBTOON))
-                                },
-                                label = { Text("Webtoon") },
-                                selected = settings.readingMode == ReadingMode.WEBTOON
-                            )
-                        }
-
-                        // Reading direction
-                        Text("Reading Direction")
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingDirection = ReadingDirection.LEFT_TO_RIGHT))
-                                },
-                                label = { Text("L → R") },
-                                selected = settings.readingDirection == ReadingDirection.LEFT_TO_RIGHT
-                            )
-                            FilterChip(
-                                onClick = {
-                                    onSettingsChanged(settings.copy(readingDirection = ReadingDirection.RIGHT_TO_LEFT))
-                                },
-                                label = { Text("R → L") },
-                                selected = settings.readingDirection == ReadingDirection.RIGHT_TO_LEFT
-                            )
-                        }
-
-                        // Volume keys navigation
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Volume Keys Navigation")
-                            Switch(
-                                checked = settings.volumeKeysForNavigation,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(volumeKeysForNavigation = it))
-                                }
-                            )
-                        }
-
-                        // Auto bookmarks
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Auto Bookmarks")
-                            Switch(
-                                checked = settings.autoBookmarks,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(autoBookmarks = it))
-                                }
-                            )
-                        }
-
-                        // Reading analytics
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Reading Analytics")
-                            Switch(
-                                checked = settings.readingAnalytics,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(readingAnalytics = it))
-                                }
-                            )
-                        }
-                    }
-                }
-            }
-
-            // Gestures
-            item {
-                SettingsCard(title = "Gestures") {
-                    Column(
-                        verticalArrangement = Arrangement.spacedBy(16.dp)
-                    ) {
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Double Tap to Zoom")
-                            Switch(
-                                checked = settings.doubleTapToZoom,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(doubleTapToZoom = it))
-                                }
-                            )
-                        }
-
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text("Pinch to Zoom")
-                            Switch(
-                                checked = settings.pinchToZoom,
-                                onCheckedChange = {
-                                    onSettingsChanged(settings.copy(pinchToZoom = it))
-                                }
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun SettingsCard(
@@ -690,142 +193,6 @@ private fun SettingsCard(
     }
 }
 
-@Composable
-fun SearchDialog(
-    searchQuery: String,
-    searchResults: List<SearchResult>,
-    currentSearchIndex: Int,
-    onDismiss: () -> Unit,
-    onSearch: (String) -> Unit,
-    onNavigateToResult: (Int) -> Unit
-) {
-    Dialog(onDismissRequest = onDismiss) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(0.8f),
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface
-            )
-        ) {
-            Column(
-                modifier = Modifier.padding(16.dp)
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = "Search in Document",
-                        style = MaterialTheme.typography.titleLarge,
-                        fontWeight = FontWeight.Medium
-                    )
-                    IconButton(onClick = onDismiss) {
-                        Icon(Icons.Default.Close, contentDescription = "Close")
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(16.dp))
-
-                var query by remember { mutableStateOf(searchQuery) }
-                val keyboardController = LocalSoftwareKeyboardController.current
-
-                OutlinedTextField(
-                    value = query,
-                    onValueChange = { query = it },
-                    label = { Text("Search text") },
-                    placeholder = { Text("Enter text to search") },
-                    modifier = Modifier.fillMaxWidth(),
-                    keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                    keyboardActions = KeyboardActions(
-                        onSearch = {
-                            onSearch(query)
-                            keyboardController?.hide()
-                        }
-                    ),
-                    trailingIcon = {
-                        Row {
-                            if (query.isNotEmpty()) {
-                                IconButton(onClick = { query = "" }) {
-                                    Icon(Icons.Default.Clear, contentDescription = "Clear")
-                                }
-                            }
-                            IconButton(
-                                onClick = {
-                                    onSearch(query)
-                                    keyboardController?.hide()
-                                }
-                            ) {
-                                Icon(Icons.Default.Search, contentDescription = "Search")
-                            }
-                        }
-                    }
-                )
-
-                if (searchResults.isNotEmpty()) {
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = "${searchResults.size} results found",
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-
-                        if (searchResults.size > 1) {
-                            Row {
-                                IconButton(
-                                    onClick = {
-                                        val newIndex = if (currentSearchIndex > 0)
-                                            currentSearchIndex - 1
-                                        else
-                                            searchResults.size - 1
-                                        onNavigateToResult(newIndex)
-                                    }
-                                ) {
-                                    Icon(Icons.Default.KeyboardArrowUp, contentDescription = "Previous")
-                                }
-                                Text(
-                                    text = "${currentSearchIndex + 1}",
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    modifier = Modifier.align(Alignment.CenterVertically)
-                                )
-                                IconButton(
-                                    onClick = {
-                                        val newIndex = if (currentSearchIndex < searchResults.size - 1)
-                                            currentSearchIndex + 1
-                                        else
-                                            0
-                                        onNavigateToResult(newIndex)
-                                    }
-                                ) {
-                                    Icon(Icons.Default.KeyboardArrowDown, contentDescription = "Next")
-                                }
-                            }
-                        }
-                    }
-
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                    LazyColumn {
-                        items(searchResults.withIndex().toList()) { (index, result) ->
-                            SearchResultItem(
-                                result = result,
-                                isSelected = index == currentSearchIndex,
-                                onClick = { onNavigateToResult(index) }
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun SearchResultItem(
