+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/reader/components/SearchDialog.kt
@@ -1,0 +1,177 @@
+package com.universalmedialibrary.ui.reader.components
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.*
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.automirrored.filled.MenuBook
+import androidx.compose.material.icons.automirrored.filled.NavigateBefore
+import androidx.compose.material.icons.automirrored.filled.NavigateNext
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import com.universalmedialibrary.ui.reader.ReadingStats
+import com.universalmedialibrary.ui.viewer.MediaViewerManager
+import com.universalmedialibrary.ui.viewer.common.ReadingDirection
+import com.universalmedialibrary.ui.viewer.common.ReadingMode
+import com.universalmedialibrary.ui.viewer.common.TextAlignment
+import com.universalmedialibrary.ui.viewer.common.ViewerSettings
+
+@Composable
+fun SearchDialog(
+    searchQuery: String,
+    searchResults: List<SearchResult>,
+    currentSearchIndex: Int,
+    onDismiss: () -> Unit,
+    onSearch: (String) -> Unit,
+    onNavigateToResult: (Int) -> Unit
+) {
+    Dialog(onDismissRequest = onDismiss) {
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .fillMaxHeight(0.8f),
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.surface
+            )
+        ) {
+            Column(
+                modifier = Modifier.padding(16.dp)
+            ) {
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Text(
+                        text = "Search in Document",
+                        style = MaterialTheme.typography.titleLarge,
+                        fontWeight = FontWeight.Medium
+                    )
+                    IconButton(onClick = onDismiss) {
+                        Icon(Icons.Default.Close, contentDescription = "Close")
+                    }
+                }
+
+                Spacer(modifier = Modifier.height(16.dp))
+
+                var query by remember { mutableStateOf(searchQuery) }
+                val keyboardController = LocalSoftwareKeyboardController.current
+
+                OutlinedTextField(
+                    value = query,
+                    onValueChange = { query = it },
+                    label = { Text("Search text") },
+                    placeholder = { Text("Enter text to search") },
+                    modifier = Modifier.fillMaxWidth(),
+                    keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
+                    keyboardActions = KeyboardActions(
+                        onSearch = {
+                            onSearch(query)
+                            keyboardController?.hide()
+                        }
+                    ),
+                    trailingIcon = {
+                        Row {
+                            if (query.isNotEmpty()) {
+                                IconButton(onClick = { query = "" }) {
+                                    Icon(Icons.Default.Clear, contentDescription = "Clear")
+                                }
+                            }
+                            IconButton(
+                                onClick = {
+                                    onSearch(query)
+                                    keyboardController?.hide()
+                                }
+                            ) {
+                                Icon(Icons.Default.Search, contentDescription = "Search")
+                            }
+                        }
+                    }
+                )
+
+                if (searchResults.isNotEmpty()) {
+                    Spacer(modifier = Modifier.height(16.dp))
+
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        horizontalArrangement = Arrangement.SpaceBetween,
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
+                        Text(
+                            text = "${searchResults.size} results found",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onSurfaceVariant
+                        )
+
+                        if (searchResults.size > 1) {
+                            Row {
+                                IconButton(
+                                    onClick = {
+                                        val newIndex = if (currentSearchIndex > 0)
+                                            currentSearchIndex - 1
+                                        else
+                                            searchResults.size - 1
+                                        onNavigateToResult(newIndex)
+                                    }
+                                ) {
+                                    Icon(Icons.Default.KeyboardArrowUp, contentDescription = "Previous")
+                                }
+                                Text(
+                                    text = "${currentSearchIndex + 1}",
+                                    style = MaterialTheme.typography.bodyMedium,
+                                    modifier = Modifier.align(Alignment.CenterVertically)
+                                )
+                                IconButton(
+                                    onClick = {
+                                        val newIndex = if (currentSearchIndex < searchResults.size - 1)
+                                            currentSearchIndex + 1
+                                        else
+                                            0
+                                        onNavigateToResult(newIndex)
+                                    }
+                                ) {
+                                    Icon(Icons.Default.KeyboardArrowDown, contentDescription = "Next")
+                                }
+                            }
+                        }
+                    }
+
+                    Spacer(modifier = Modifier.height(8.dp))
+
+                    LazyColumn {
+                        items(searchResults.withIndex().toList()) { (index, result) ->
+                            SearchResultItem(
+                                result = result,
+                                isSelected = index == currentSearchIndex,
+                                onClick = { onNavigateToResult(index) }
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
